increase form label widths
In the registration template, both the “Terms of Use” and “Remember Me”
checkbox labels were subject to the universal `.affwp-form label` width
of 200px. That’s cool when the labels are to the left of a text input
field but made bad use of open space (to the right) for label text that
has nothing but a checkbox to the left.
http://glui.me/?i=9hbrin2qzkcjzbl/2014-05-05_at_1.00_AM.png/

There was no markup in place to target these labels so I added the
classes. I combined the two new CSS selectors with the selector for a
label located in the dashboard template that has already addressed this
exact issue (referral notifications checkbox label).