Modify return type of affwp_add_referral()
affwp_add_referral() becomes a bit more helpful as a wrapper function for `affiliate_wp()->referrals->add( $args )` if it matches the return value.

Now, we can use the returned value of the referral to actually do something with the referral that was just added.  Woot!

All the falsey/truthy checks in core (maybe 1?) will still work fine, as 0 evaluates to false, and any other positive integer evaluates to true.